refactor: move recent DwC-A extensions into a subclass
* Allow archive.rb to omit extension-specific logic for describing extension
  files and generating data
* Adds metadata for DNADerivedData fields and files
