+++ spec/lib/darwin_core/extensions/vernacular_names.rb
@@ -2,7 +2,7 @@
 
 require "spec_helper"
 
-describe DarwinCore::VernacularName do
+describe DarwinCore::Extensions::VernacularNames do
   let( :taxon_name_without_lexicon ) do
     taxon_name = create :taxon_name
     TaxonName.where( id: taxon_name.id ).update_all( "lexicon = null" )
@@ -13,7 +13,7 @@ describe DarwinCore::VernacularName do
 
   describe "language" do
     it "should be und if lexicon is null" do
-      expect( DarwinCore::VernacularName.adapt( taxon_name_without_lexicon ).language ).to eq "und"
+      expect( DarwinCore::Extensions::VernacularNames.adapt( taxon_name_without_lexicon ).language ).to eq "und"
     end
   end
 
@@ -22,7 +22,7 @@ describe DarwinCore::VernacularName do
       taxon_name = taxon_name_without_lexicon
       expect( taxon_name.taxon ).to be_is_active
       expect( taxon_name.lexicon ).to be_blank
-      paths = DarwinCore::VernacularName.data( core: DarwinCore::Cores::TAXON )
+      paths = DarwinCore::Extensions::VernacularNames.data( core: DarwinCore::Cores::TAXON )
       puts "paths: #{paths}"
       row = CSV.read( paths.first, headers: true ).first
       expect( row["vernacularName"] ).to eq taxon_name.name
